C-Multi-Layered Content Integrity Verification/deepshield_watermark/pyproject.toml
[tool.poetry]
name = "deepshield-watermark"
version = "0.1.0"
description = ""
authors = ["Mohamed Njeh <njeh@iabg.de>", "Vaghasiya Nehal <vaghasiya@iabg.de>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10"
pillow = "^11.0.0"
numpy = "^2.2.1"
pywavelets = "^1.8.0"
pytest = "^8.3.4"
scipy = "^1.15.1"
scikit-image = "^0.25.0"
pycryptodome = "^3.21.0"
torchvision = "^0.20.1"
pycocotools = "^2.0.8"
opencv-python = "^4.11.0.86"
tqdm = "^4.67.1"
openai = "^1.61.1"
requests = "^2.32.3"
pyyaml = "^6.0.2"
rich = "^13.9.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"
pylint = "^3.3.2"
flake8 = "^7.1.1"

[tool.black]
line-length = 100
target-version = ['py310']


[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
